Fix frame indices

+++ Fourier Synth.swiftpm/Sources/Model/SynthesizerContext.swift
@@ -7,6 +7,6 @@
 
 /// The state of the synthesizer relevant for rendering individual nodes.
 struct SynthesizerContext: Hashable, Codable, Sendable {
-    var frame: Double = 0
+    var frame: Int = 0
     var sampleRate: Double = 0
 }
